move perforamnce tests to testing
